+++ CHANGES.md
@@ -13,6 +13,7 @@
   + `Graphql.exec` now takes a `request`
   + `Contributions.fetch` has been replaced by `Contributions.request` that builds a `request`
 - Add a `~user:User.t` parameter to `Contributions.request` and `Contributions.of_json` (#14, @gpetiot)
+- `Contributions.of_json` now returns a result type (#20, @gpetiot)
 
 ## 0.2.0